separate tests for important functions from tests for parallelized implementations
